After reconfiguring architecture, all old modules have been marked as such. Each module is being integrated into the sprint 3 implementation one at a time. I have integrated login with Firebase Authentication.
