+++ client/package.json
@@ -16,6 +16,7 @@
     "@angular/common": "^18.2.0",
     "@angular/compiler": "^18.2.0",
     "@angular/core": "^18.2.0",
+    "@angular/fire": "^18.0.1",
     "@angular/forms": "^18.2.0",
     "@angular/material": "^18.2.7",
     "@angular/platform-browser": "^18.2.0",
