+++ client/src/app/app-routing.module.ts
@@ -1,20 +1,20 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AppComponent } from './app.component';
-import { AccountComponent } from './basicFunctionality/account/account.component';
-import { ChangePasswordComponent } from './basicFunctionality/change-password/change-password.component';
-import { ForgetPasswordComponent } from './basicFunctionality/forget-password/forget-password.component';
-import { HomeComponent } from './basicFunctionality/home/home.component';
+import { AccountComponent } from './basicFunctionality/account-old/account.component';
+import { ChangePasswordComponent } from './basicFunctionality/change-password-old/change-password.component';
+import { ForgetPasswordComponent } from './basicFunctionality/forget-password-old/forget-password.component';
+import { HomeComponent } from './basicFunctionality/home-old/home.component';
 import { LoginComponent } from './basicFunctionality/login/login.component';
-import { RegisterComponent } from './basicFunctionality/register/register.component';
-import { ResetPasswordComponent } from './basicFunctionality/reset-password/reset-password.component';
-import { RoleComponent } from './basicFunctionality/role/role.component';
-import { UsersComponent } from './basicFunctionality/users/users.component';
-import { NavbarComponent } from './basicFunctionality/navbar/navbar.component';
+import { RegisterComponent } from './basicFunctionality/register-old/register.component';
+import { ResetPasswordComponent } from './basicFunctionality/reset-password-old/reset-password.component';
+import { RoleComponent } from './basicFunctionality/role-old/role.component';
+import { UsersComponent } from './basicFunctionality/users-old/users.component';
+import { NavbarComponent } from './basicFunctionality/navbar-old/navbar.component';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
-import { ChartComponent } from './basicFunctionality/chart/chart.component';
-import { UserProfileComponent } from './basicFunctionality/user-profile/user-profile.component';
-import { AccountProfileComponent } from './basicFunctionality/account-profile/account-profile.component';
+import { ChartComponent } from './basicFunctionality/chart-old/chart.component';
+import { UserProfileComponent } from './basicFunctionality/user-profile-old/user-profile.component';
+import { AccountProfileComponent } from './basicFunctionality/account-profile-old/account-profile.component';
 import { CalendarComponent } from './basicFunctionality/calendar/calendar.component';
 import { DashboardComponent } from './basicFunctionality/dashboard/dashboard.component';
