+++ client/src/app/basicFunctionality/account-profile-old/account-profile.component.ts
@@ -8,27 +8,32 @@ import { MatMenuModule } from '@angular/material/menu';
 import { CommonModule } from '@angular/common';
 import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { ApiService } from '../../services/api.service';
-import { UserModel } from '../account/account.model';
+import { UserModel } from '../account-old/account.model';
+import { ChartModel } from '../chart-old/chart.model';
+import { ChartService } from '../../services/chart.service';
 
 @Component({
-  selector: 'app-user-profile',
-  templateUrl: './user-profile.component.html',
-  styleUrl: './user-profile.component.scss'
+  selector: 'app-account-profile',
+  templateUrl: './account-profile.component.html',
+  styleUrl: './account-profile.component.scss'
 })
-export class UserProfileComponent   implements OnInit {
+export class AccountProfileComponent implements OnInit {
 
   public profileDetail: UserModel = new UserModel();
   public reset: UserModel = new UserModel();
 
+  public chartDetail: ChartModel = new ChartModel();
 
-  constructor(private api: ApiService, private router: Router) { }
+
+  constructor(private api: ApiService, private router: Router, private chart: ChartService,) { }
 
   ngOnInit(): void {
+
     if (typeof window !== 'undefined') {
-      if (localStorage.getItem('profile') != null) {
-        const temp = localStorage.getItem('profile');
+      if (localStorage.getItem('a-profile') != null) {
+        const temp = localStorage.getItem('a-profile');
         if (temp != null) {
-          this.profileDetail = JSON.parse(temp);
+          this.chartDetail = JSON.parse(temp);
         }
       }
     }
