+++ client/src/app/basicFunctionality/home-old/home.component.ts
@@ -1,4 +1,4 @@
-import { Component, inject, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatIconModule } from '@angular/material/icon';
@@ -8,16 +8,14 @@ import { MatMenuModule } from '@angular/material/menu';
 import { CommonModule } from '@angular/common';
 import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { ApiService } from '../../services/api.service';
-import { UserModel } from '../account/account.model';
-import { CalendarModule, DateAdapter } from 'angular-calendar';
-import { adapterFactory } from 'angular-calendar/date-adapters/date-fns';
+import { UserModel } from '../account-old/account.model';
 
 @Component({
-  selector: 'app-navbar',
-  templateUrl: './navbar.component.html',
-  styleUrl: './navbar.component.scss'
+  selector: 'app-home',
+  templateUrl: './home.component.html',
+  styleUrl: './home.component.scss'
 })
-export class NavbarComponent implements OnInit {
+export class HomeComponent implements OnInit{
 
   public userDetail: UserModel = new UserModel();
   public reset: UserModel = new UserModel();
@@ -59,19 +57,8 @@ export class NavbarComponent implements OnInit {
 
   }
 
-  logout = () => {
-    if (localStorage.getItem('user') != null) {
-      localStorage.removeItem('user');
-    }
-    this.userDetail = this.reset;
-
-    this.router.navigate(['/login']);
-  };
 
   getRole(): number {
     return this.api.getRole();
   }
-
-
-
 }
