+++ client/src/app/basicFunctionality/login/login.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, inject } from '@angular/core';
-import { UserModel } from '../account/account.model';
+import { UserModel } from '../account-old/account.model';
 import { ApiService } from '../../services/api.service';
 import { Router, RouterModule, RouterLink } from '@angular/router';
 import {
@@ -9,6 +9,8 @@ import {
   Validators,
 } from '@angular/forms';
 import { HttpClient } from '@angular/common/http';
+import { Auth, signInWithEmailAndPassword } from '@angular/fire/auth'; //While it says email, we are going to just have the user provide the username.
+                                                                       //and then we can just append '@midas.com' to the username to make it an email.
 
 @Component({
   selector: 'app-login',
@@ -18,7 +20,7 @@ import { HttpClient } from '@angular/common/http';
 export class LoginComponent implements OnInit {
 
   formValue !: FormGroup;
-  userModelObj : UserModel = new UserModel();
+  private auth = inject(Auth);
   userData !: any;
   passwordHide: boolean = true;
   private userKey = 'userKey';
@@ -26,7 +28,15 @@ export class LoginComponent implements OnInit {
   constructor(private formbuilder: FormBuilder, private api : ApiService, private router: Router, private http: HttpClient){}
 
   login() {
-        this.api.login(this.formValue.value.username, this.formValue.value.password);
+    signInWithEmailAndPassword(this.auth, this.formValue.value.username + '@midas.com', this.formValue.value.password)
+        .then((userCredential) => {
+          const user = userCredential.user;
+        })
+        .catch((error) => {
+          console.error(error);
+        })
+        //this.api.login(this.formValue.value.username, this.formValue.value.password);
+
   }
