+++ client/src/app/basicFunctionality/navbar-old/navbar.component.scss
