+++ client/src/app/basicFunctionality/user-profile-old/user-profile.component.ts
@@ -8,32 +8,27 @@ import { MatMenuModule } from '@angular/material/menu';
 import { CommonModule } from '@angular/common';
 import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { ApiService } from '../../services/api.service';
-import { UserModel } from '../account/account.model';
-import { ChartModel } from '../chart/chart.model';
-import { ChartService } from '../../services/chart.service';
+import { UserModel } from '../account-old/account.model';
 
 @Component({
-  selector: 'app-account-profile',
-  templateUrl: './account-profile.component.html',
-  styleUrl: './account-profile.component.scss'
+  selector: 'app-user-profile',
+  templateUrl: './user-profile.component.html',
+  styleUrl: './user-profile.component.scss'
 })
-export class AccountProfileComponent implements OnInit {
+export class UserProfileComponent   implements OnInit {
 
   public profileDetail: UserModel = new UserModel();
   public reset: UserModel = new UserModel();
 
-  public chartDetail: ChartModel = new ChartModel();
 
-
-  constructor(private api: ApiService, private router: Router, private chart: ChartService,) { }
+  constructor(private api: ApiService, private router: Router) { }
 
   ngOnInit(): void {
-
     if (typeof window !== 'undefined') {
-      if (localStorage.getItem('a-profile') != null) {
-        const temp = localStorage.getItem('a-profile');
+      if (localStorage.getItem('profile') != null) {
+        const temp = localStorage.getItem('profile');
         if (temp != null) {
-          this.chartDetail = JSON.parse(temp);
+          this.profileDetail = JSON.parse(temp);
         }
       }
     }
