+++ client/src/app/basicFunctionality/users-old/users.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatIconModule } from '@angular/material/icon';
@@ -8,31 +8,22 @@ import { MatMenuModule } from '@angular/material/menu';
 import { CommonModule } from '@angular/common';
 import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { ApiService } from '../../services/api.service';
-import { UserModel } from '../account/account.model';
+import { UserModel } from '../account-old/account.model';
+import { UserService } from '../../services/user.service';
 
 @Component({
-  selector: 'app-home',
-  templateUrl: './home.component.html',
-  styleUrl: './home.component.scss'
+  selector: 'app-users',
+  templateUrl: './users.component.html',
+  styleUrl: './users.component.scss'
 })
-export class HomeComponent implements OnInit{
+export class UsersComponent {
 
   public userDetail: UserModel = new UserModel();
   public reset: UserModel = new UserModel();
 
-
+  userService = inject(UserService);
   constructor(private api: ApiService, private router: Router) { }
 
-  ngOnInit(): void {
-    if (typeof window !== 'undefined') {
-      if (localStorage.getItem('user') != null) {
-        const temp = localStorage.getItem('user');
-        if (temp != null) {
-          this.userDetail = JSON.parse(temp);
-        }
-      }
-    }
-  }
 
   isLoggedIn = (): boolean => {
 
@@ -57,6 +48,14 @@ export class HomeComponent implements OnInit{
 
   }
 
+  logout = () => {
+    if (localStorage.getItem('user') != null) {
+      localStorage.removeItem('user');
+    }
+    this.userDetail = this.reset;
+
+    this.router.navigate(['/login']);
+  };
 
   getRole(): number {
     return this.api.getRole();
